Renaming field components
