+++ admin/src/fields.js
@@ -1,29 +1,29 @@
 module.exports = {
-	boolean:          require('../../fields/types/boolean/boolean'),
-	date:             require('../../fields/types/date/date'),
-	datetime:         require('../../fields/types/datetime/datetime'),
-	email:            require('../../fields/types/email/email'),
-	location:         require('../../fields/types/location/location'),
-	markdown:         require('../../fields/types/markdown/markdown'),
-	name:             require('../../fields/types/name/name'),
-	money:            require('../../fields/types/money/money'),
-	password:         require('../../fields/types/password/password'),
-	select:           require('../../fields/types/select/select'),
-	text:             require('../../fields/types/text/text'),
-	key:              require('../../fields/types/key/key'),
-	textarea:         require('../../fields/types/textarea/textarea'),
-	url:              require('../../fields/types/url/url'),
-	embedly:          require('../../fields/types/embedly/embedly'),
-	color:            require('../../fields/types/color/color'),
-	relationship:     require('../../fields/types/relationship/relationship'),
-	s3file:           require('../../fields/types/s3file/s3file'),
-	azurefile:        require('../../fields/types/azurefile/azurefile'),
-	localfile:  	  require('../../fields/types/localfile/localfile'),
-	textarray:        require('../../fields/types/textarray/textarray'),
-	numberarray:      require('../../fields/types/numberarray/numberarray'),
-	code:             require('../../fields/types/code/code'),
-	number:           require('../../fields/types/number/number'),
-	html:             require('../../fields/types/html/html'),
-	cloudinaryimage:  require('../../fields/types/cloudinaryimage/cloudinaryimage'),
-	cloudinaryimages: require('../../fields/types/cloudinaryimages/cloudinaryimages')
+	boolean:          require('../../fields/types/boolean/BooleanField'),
+	date:             require('../../fields/types/date/DateField'),
+	datetime:         require('../../fields/types/datetime/DateTimeField'),
+	email:            require('../../fields/types/email/EmailField'),
+	location:         require('../../fields/types/location/LocationField'),
+	markdown:         require('../../fields/types/markdown/MarkdownField'),
+	name:             require('../../fields/types/name/NameField'),
+	money:            require('../../fields/types/money/MoneyField'),
+	password:         require('../../fields/types/password/PasswordField'),
+	select:           require('../../fields/types/select/SelectField'),
+	text:             require('../../fields/types/text/TextField'),
+	key:              require('../../fields/types/key/KeyField'),
+	textarea:         require('../../fields/types/textarea/TextareaField'),
+	url:              require('../../fields/types/url/UrlField'),
+	embedly:          require('../../fields/types/embedly/EmbedlyField'),
+	color:            require('../../fields/types/color/ColorField'),
+	relationship:     require('../../fields/types/relationship/RelationshipField'),
+	s3file:           require('../../fields/types/s3file/S3FileField'),
+	azurefile:        require('../../fields/types/azurefile/AzureFileField'),
+	localfile:  	  require('../../fields/types/localfile/LocalFileField'),
+	textarray:        require('../../fields/types/textarray/TextArrayField'),
+	numberarray:      require('../../fields/types/numberarray/NumberArrayField'),
+	code:             require('../../fields/types/code/CodeField'),
+	number:           require('../../fields/types/number/NumberField'),
+	html:             require('../../fields/types/html/HtmlField'),
+	cloudinaryimage:  require('../../fields/types/cloudinaryimage/CloudinaryImageField'),
+	cloudinaryimages: require('../../fields/types/cloudinaryimages/CloudinaryImagesField')
 };
