+++ fields/types/azurefile/AzureFileField.js
@@ -1,0 +1,1 @@
+module.exports = require('../localfile/LocalFileField');
