--- fields/types/azurefile/azurefile.js
@@ -1,1 +1,0 @@
-module.exports = require('../localfile/localfile');
