+++ fields/types/cloudinaryimage/CloudinaryImageField.js
@@ -1,8 +1,8 @@
-var _      = require('underscore'),
-	$      = require('jquery'),
-	React  = require('react'),
-	Field  = require('../field'),
-	Note   = require('../../components/note'),
+var _ = require('underscore'),
+	$ = require('jquery'),
+	React = require('react'),
+	Field = require('../field'),
+	Note = require('../../components/Note'),
 	Select = require('react-select');
 
 var SUPPORTED_TYPES = ['image/gif', 'image/png', 'image/jpeg', 'image/bmp', 'image/x-icon', 'application/pdf', 'image/x-tiff', 'image/x-tiff', 'application/postscript', 'image/vnd.adobe.photoshop'];
