+++ fields/types/cloudinaryimages/CloudinaryImagesField.js
@@ -1,5 +1,5 @@
-var _     = require('underscore'),
-	$     = require('jquery'),
+var _ = require('underscore'),
+	$ = require('jquery'),
 	React = require('react'),
 	Field = require('../field');
