+++ fields/types/embedly/EmbedlyField.js
@@ -1,5 +1,5 @@
 var React = require('react'),
-	Field = require('../field');
+	Field = require('../Field');
 
 module.exports = Field.create({
