+++ fields/types/localfile/LocalFileField.js
@@ -1,8 +1,8 @@
-var _      = require('underscore'),
-	$      = require('jquery'),
-	React  = require('react'),
-	Field  = require('../field'),
-	Note   = require('../../components/note'),
+var _ = require('underscore'),
+	$ = require('jquery'),
+	React = require('react'),
+	Field = require('../Field'),
+	Note = require('../../components/Note'),
 	Select = require('react-select');
 
 module.exports = Field.create({
