+++ fields/types/name/NameField.js
@@ -1,6 +1,6 @@
 var React = require('react'),
-	Field = require('../field'),
-	Note = require('../../components/note');
+	Field = require('../Field'),
+	Note = require('../../components/Note');
 
 module.exports = Field.create({
