+++ fields/types/password/PasswordField.js
@@ -1,7 +1,7 @@
 var _ = require('underscore'),
 	React = require('react'),
-	Field = require('../field'),
-	Note = require('../../components/note');
+	Field = require('../Field'),
+	Note = require('../../components/Note');
 
 module.exports = Field.create({
