+++ fields/types/s3file/S3FileField.js
@@ -1,0 +1,1 @@
+module.exports = require('../localfile/LocalFileField');
