--- fields/types/s3file/s3file.js
@@ -1,1 +1,0 @@
-module.exports = require('../localfile/localfile');
