+++ fields/types/url/UrlField.js
@@ -1,4 +1,4 @@
 var React = require('react'),
-	Field = require('../field');
+	Field = require('../Field');
 
 module.exports = Field.create();
